Resolve merge conflicts in test_buffer_traditional.py caused by upstream
changes that migrated from duthost.shell() to dut_asic.run_redis_cmd()
and introduced extract_profile_name(). Preserves the empty profile_in_pg
guard from this branch while adopting upstream's API and style changes.

+++ tests/copp/copp_utils.py
@@ -7,6 +7,9 @@
 import re
 import logging
 import json
+import ipaddress
+import ast
+import random
 
 from tests.common.config_reload import config_reload
 
@@ -38,7 +41,7 @@ _CONFIG_DB = "/etc/sonic/config_db.json"
 _TEMP_CONFIG_DB = "/home/admin/config_db_copp_backup.json"
 
 
-def limit_policer(dut, pps_limit, nn_target_namespace):
+def limit_policer(dut, pps_limit, nn_target_namespace, neighbor_miss_trap_supported):
     """
         Updates the COPP configuration in the SWSS container to respect a given rate limit.
 
@@ -63,11 +66,13 @@ def limit_policer(dut, pps_limit, nn_target_namespace):
         config_format = "config_db"
 
     dut.script(
-        cmd="{} {} {} {} {}".format(_UPDATE_COPP_SCRIPT,
-                                    pps_limit,
-                                    _BASE_COPP_CONFIG,
-                                    _TEMP_COPP_CONFIG,
-                                    config_format)
+        cmd="{} {} {} {} {} {} {}".format(_UPDATE_COPP_SCRIPT,
+                                          pps_limit,
+                                          _BASE_COPP_CONFIG,
+                                          _TEMP_COPP_CONFIG,
+                                          config_format,
+                                          dut.facts["asic_type"],
+                                          neighbor_miss_trap_supported)
     )
 
     if config_format == "app_db":
@@ -80,6 +85,7 @@ def limit_policer(dut, pps_limit, nn_target_namespace):
         dut.command("docker cp {} {}".format(_TEMP_COPP_CONFIG, swss_docker_name + _SWSS_COPP_TEMPLATE))
     else:
         dut.command("cp {} {}".format(_TEMP_COPP_CONFIG, _DEFAULT_COPP_TEMPLATE))
+
 
 def restore_policer(dut, nn_target_namespace):
     """
@@ -114,7 +120,9 @@ def configure_ptf(ptf, test_params, is_backend_topology=False):
 
     ptf.script(cmd=_REMOVE_IP_SCRIPT)
     if is_backend_topology:
-        ip_command = "ip address add %s/31 dev \"eth%s.%s\"" % (test_params.myip, test_params.nn_target_port, test_params.nn_target_vlanid)
+        ip_command = "ip address add %s/31 dev \"eth%s.%s\"" % (test_params.myip,
+                                                                test_params.nn_target_port,
+                                                                test_params.nn_target_vlanid)
     else:
         ip_command = "ip address add %s/31 dev eth%s" % (test_params.myip, test_params.nn_target_port)
 
@@ -129,6 +137,7 @@ def configure_ptf(ptf, test_params, is_backend_topology=False):
     ptf.template(src=_PTF_NN_TEMPLATE, dest=_PTF_NN_DEST)
 
     ptf.supervisorctl(name="ptf_nn_agent", state="restarted")
+
 
 def restore_ptf(ptf):
     """
@@ -149,6 +158,7 @@ def restore_ptf(ptf):
     ptf.template(src=_PTF_NN_TEMPLATE, dest=_PTF_NN_DEST)
 
     ptf.supervisorctl(name="ptf_nn_agent", state="restarted")
+
 
 def configure_syncd(dut, nn_target_port, nn_target_interface, nn_target_namespace, nn_target_vlanid, swap_syncd, creds):
     """
@@ -188,14 +198,39 @@ def configure_syncd(dut, nn_target_port, nn_target_interface, nn_target_namespac
     dut.command("docker exec {} supervisorctl reread".format(syncd_docker_name))
     dut.command("docker exec {} supervisorctl update".format(syncd_docker_name))
 
+
 def restore_syncd(dut, nn_target_namespace):
     asichost = dut.asic_instance_from_namespace(nn_target_namespace)
-
     syncd_docker_name = asichost.get_docker_name("syncd")
+    asichost.stop_service("syncd")
+    asichost.delete_container(syncd_docker_name)
 
-    dut.command("docker exec {} rm -rf /etc/supervisor/conf.d/{}".format(syncd_docker_name, _SYNCD_NN_FILE))
-    dut.command("docker exec {} supervisorctl reread".format(syncd_docker_name))
-    dut.command("docker exec {} supervisorctl update".format(syncd_docker_name))
+
+def _install_nano_bookworm(dut, creds, syncd_docker_name):
+    output = dut.command("docker exec {} bash -c '[ -d /usr/include/nanomsg ] || \
+        echo copp'".format(syncd_docker_name))
+
+    if output["stdout"] == "copp":
+        http_proxy = creds.get('proxy_env', {}).get('http_proxy', '')
+        https_proxy = creds.get('proxy_env', {}).get('https_proxy', '')
+        # Change the permission of /tmp to 1777 to workaround issue sonic-net/sonic-buildimage#16034
+        cmd = '''docker exec -e http_proxy={} -e https_proxy={} {} bash -c " \
+                mkdir -p /var/tmp_build \
+                && rm -rf /var/lib/apt/lists/* \
+                && apt-get update \
+                && apt-get install -y python3-pip build-essential libssl-dev libffi-dev \
+                python3-dev python3-setuptools wget libnanomsg-dev python-is-python3 \
+                && TMPDIR=/var/tmp_build pip3 install --no-cache-dir cffi==1.16.0 \
+                && TMPDIR=/var/tmp_build pip3 install --no-cache-dir nnpy \
+                && rm -rf /var/tmp_build \
+                && mkdir -p /opt && cd /opt && wget \
+                https://raw.githubusercontent.com/p4lang/ptf/master/ptf_nn/ptf_nn_agent.py \
+                && mkdir ptf && cd ptf && wget \
+                https://raw.githubusercontent.com/p4lang/ptf/master/src/ptf/afpacket.py && touch __init__.py \
+                && apt-get -y purge build-essential libssl-dev libffi-dev python3-dev \
+                python3-setuptools wget \
+                " '''.format(http_proxy, https_proxy, syncd_docker_name)
+        dut.command(cmd)
 
 
 def _install_nano(dut, creds,  syncd_docker_name):
@@ -206,46 +241,59 @@ def _install_nano(dut, creds,  syncd_docker_name):
             dut (SonicHost): The target device.
             creds (dict): Credential information according to the dut inventory
     """
-    output = dut.command("docker exec {} bash -c '[ -d /usr/local/include/nanomsg ] && [ -d /opt/ptf ] || echo copp'".format(syncd_docker_name))
+
+    if "bookworm" in dut.shell("docker exec {} grep VERSION_CODENAME /etc/os-release"
+                               .format(syncd_docker_name))['stdout'].lower():
+        _install_nano_bookworm(dut, creds, syncd_docker_name)
+        return
+
+    if dut.facts["asic_type"] == "cisco-8000":
+        output = dut.command("docker exec {} bash -c '[ -d /usr/local/include/nanomsg ] || \
+            echo copp'".format(syncd_docker_name))
+    else:
+        output = dut.command("docker exec {} bash -c '[ -d /usr/local/include/nanomsg ] && [ -d /opt/ptf ] || \
+            echo copp'".format(syncd_docker_name))
 
     if output["stdout"] == "copp":
         http_proxy = creds.get('proxy_env', {}).get('http_proxy', '')
         https_proxy = creds.get('proxy_env', {}).get('https_proxy', '')
         check_cmd = "docker exec -i {} bash -c 'cat /etc/os-release'".format(syncd_docker_name)
-
+        # Change the permission of /tmp to 1777 to workaround issue sonic-net/sonic-buildimage#16034
         if "bullseye" in dut.shell(check_cmd)['stdout'].lower():
             cmd = '''docker exec -e http_proxy={} -e https_proxy={} {} bash -c " \
-                    rm -rf /var/lib/apt/lists/* \
+                    chmod 1777 /tmp \
+                    && rm -rf /var/lib/apt/lists/* \
                     && apt-get update \
-                    && apt-get install -y python3-pip build-essential libssl-dev libffi-dev python3-dev python-setuptools wget cmake python-is-python3 \
+                    && apt-get install -y python3-pip build-essential libssl-dev libffi-dev \
+                    python3-dev python-setuptools wget cmake python-is-python3 \
                     && wget https://github.com/nanomsg/nanomsg/archive/1.0.0.tar.gz \
                     && tar xzf 1.0.0.tar.gz && cd nanomsg-1.0.0 \
                     && mkdir -p build && cmake . && make install && ldconfig && cd .. && rm -rf nanomsg-1.0.0 \
                     && rm -f 1.0.0.tar.gz && pip3 install cffi && pip3 install --upgrade cffi && pip3 install nnpy \
-                    && mkdir -p /opt && cd /opt && wget https://raw.githubusercontent.com/p4lang/ptf/master/ptf_nn/ptf_nn_agent.py \
-                    && mkdir ptf && cd ptf && wget https://raw.githubusercontent.com/p4lang/ptf/master/src/ptf/afpacket.py && touch __init__.py \
+                    && mkdir -p /opt && cd /opt && wget \
+                    https://raw.githubusercontent.com/p4lang/ptf/master/ptf_nn/ptf_nn_agent.py \
+                    && mkdir ptf && cd ptf && wget \
+                    https://raw.githubusercontent.com/p4lang/ptf/master/src/ptf/afpacket.py && touch __init__.py \
                     " '''.format(http_proxy, https_proxy, syncd_docker_name)
         else:
             cmd = '''docker exec -e http_proxy={} -e https_proxy={} {} bash -c " \
-                    rm -rf /var/lib/apt/lists/* \
+                    chmod 1777 /tmp \
+                    && rm -rf /var/lib/apt/lists/* \
                     && apt-get update \
-                    && apt-get install -y python-pip build-essential libssl-dev libffi-dev python-dev python-setuptools wget cmake \
+                    && apt-get install -y python-pip build-essential libssl-dev libffi-dev \
+                    python-dev python-setuptools wget cmake \
                     && wget https://github.com/nanomsg/nanomsg/archive/1.0.0.tar.gz \
                     && tar xzf 1.0.0.tar.gz && cd nanomsg-1.0.0 \
                     && mkdir -p build && cmake . && make install && ldconfig && cd .. && rm -rf nanomsg-1.0.0 \
-                    && rm -f 1.0.0.tar.gz && pip2 install cffi==1.7.0 && pip2 install --upgrade cffi==1.7.0 && pip2 install nnpy \
-                    && mkdir -p /opt && cd /opt && wget https://raw.githubusercontent.com/p4lang/ptf/master/ptf_nn/ptf_nn_agent.py \
-                    && mkdir ptf && cd ptf && wget https://raw.githubusercontent.com/p4lang/ptf/master/src/ptf/afpacket.py && touch __init__.py \
+                    && rm -f 1.0.0.tar.gz && pip2 install cffi==1.7.0 && pip2 install --upgrade \
+                    cffi==1.7.0 && pip2 install nnpy \
+                    && mkdir -p /opt && cd /opt && wget \
+                    https://raw.githubusercontent.com/p4lang/ptf/master/ptf_nn/ptf_nn_agent.py \
+                    && mkdir ptf && cd ptf && wget \
+                    https://raw.githubusercontent.com/p4lang/ptf/master/src/ptf/afpacket.py && touch __init__.py \
                     " '''.format(http_proxy, https_proxy, syncd_docker_name)
+        dut.command(cmd)
 
-        try:
-            # Stop bgp sessions
-            dut.command("sudo config feature autorestart bgp disabled")
-            dut.command("sudo config feature state bgp disabled")
-            dut.command(cmd)
-        finally:
-            dut.command("sudo config feature state bgp enabled")
-            dut.command("sudo config feature autorestart bgp enabled")
 
 def _map_port_number_to_interface(dut, nn_target_port):
     """
@@ -254,6 +302,7 @@ def _map_port_number_to_interface(dut, nn_target_port):
 
     interfaces = dut.command("portstat")["stdout_lines"][2:]
     return interfaces[nn_target_port].split()[0]
+
 
 def _get_http_and_https_proxy_ip(creds):
     """
@@ -349,7 +398,7 @@ def restore_config_db(dut):
     """
     dut.command("sudo cp {} {}".format(_TEMP_CONFIG_DB, _CONFIG_DB))
     dut.command("sudo rm -f {}".format(_TEMP_CONFIG_DB))
-    config_reload(dut)
+    config_reload(dut, safe_reload=True, check_intf_up_ports=True)
 
 
 def uninstall_trap(dut, feature_name, trap_id):
@@ -361,7 +410,10 @@ def uninstall_trap(dut, feature_name, trap_id):
         feature_name (str): feature name corresponding to the trap
         trap_id (str): trap id
     """
-    disable_feature_entry(dut, feature_name)
+    feature_list, _ = dut.get_feature_status()
+    if feature_name in feature_list.keys():
+        disable_feature_entry(dut, feature_name)
+
     configure_always_enabled_for_trap(dut, trap_id, "false")
 
 
@@ -388,3 +440,195 @@ def install_trap(dut, feature_name):
         feature_name (str): feature name
     """
     enable_feature_entry(dut, feature_name)
+
+
+def get_vlan_ip(duthost, ip_version):
+    """
+    @Summary: Get an IP on the Vlan subnet
+    @param duthost: Ansible host instance of the device
+    @return: Return a vlan IP, e.g., "192.168.0.2"
+    """
+
+    mg_facts = duthost.minigraph_facts(
+        host=duthost.hostname)['ansible_facts']
+    mg_vlans = mg_facts['minigraph_vlans']
+
+    if not mg_vlans:
+        return None
+
+    mg_vlan_intfs = mg_facts['minigraph_vlan_interfaces']
+
+    if ip_version == "4":
+        vlan_subnet = ipaddress.ip_network(mg_vlan_intfs[0]['subnet'])
+    else:
+        vlan_subnet = ipaddress.ip_network(mg_vlan_intfs[-1]['subnet'])
+
+    ip_addr = str(vlan_subnet[2])
+    return ip_addr
+
+
+def get_lo_ipv4(duthost):
+
+    loopback_ip = None
+    mg_facts = duthost.minigraph_facts(
+        host=duthost.hostname)['ansible_facts']
+
+    for intf in mg_facts["minigraph_lo_interfaces"]:
+        if ipaddress.ip_address(intf["addr"]).version == 4:
+            loopback_ip = intf["addr"]
+            break
+
+    return loopback_ip
+
+
+def get_copp_trap_capabilities(duthost):
+    """
+    Fetches supported trap IDs from COPP_TRAP_CAPABILITY_TABLE in STATE_DB and returns them as a list.
+    Args:
+        duthost (SonicHost): The target device.
+    Returns:
+        list: A list of supported trap IDs.
+    """
+
+    trap_ids = duthost.shell("sonic-db-cli STATE_DB HGET 'COPP_TRAP_CAPABILITY_TABLE|traps' trap_ids")['stdout']
+    return trap_ids.split(",")
+
+
+def parse_show_copp_configuration(duthost, namespace):
+    """
+    Parses the output of the `show copp configuration` command into a structured dictionary.
+    Args:
+        duthost (SonicHost): The target device.
+    Returns:
+        dict: A dictionary mapping trap IDs to their configuration details.
+    """
+    command = "show copp configuration"
+    if namespace is not None:
+        command = namespace.ns_arg + ' ' + command
+    copp_config_output = duthost.shell(command)["stdout"]
+    copp_config_lines = copp_config_output.splitlines()
+
+    # Parse the command output into a structured format
+    copp_config_data = {}
+    for line in copp_config_lines[1:]:  # Skip the header line
+        fields = line.split()
+        if len(fields) >= 8:
+            trap_id = fields[0]
+            copp_config_data[trap_id] = {
+                "trap_group": fields[1],
+                "trap_action": fields[2],
+                "cbs": fields[3],
+                "cir": fields[4],
+                "meter_type": fields[5],
+                "mode": fields[6],
+                "hw_status": fields[7]
+            }
+
+    return copp_config_data
+
+
+def is_trap_installed(duthost, trap_id, namespace=None):
+    """
+    Checks if a specific trap is installed by parsing the output of `show copp configuration`.
+    Args:
+        dut (SonicHost): The target device
+        trap_id: The trap ID to check.
+    Returns:
+        bool: True if the trap is installed, False otherwise.
+    """
+
+    output = parse_show_copp_configuration(duthost, namespace)
+    assert trap_id in output, f"Trap {trap_id} not found in the configuration"
+    assert "hw_status" in output[trap_id], f"hw_status not found for trap {trap_id}"
+
+    return output[trap_id]["hw_status"] == "installed"
+
+
+def is_trap_uninstalled(duthost, trap_id):
+    """
+    Checks if a specific trap is uninstalled by parsing the output of `show copp configuration`.
+    Args:
+        dut (SonicHost): The target device
+        trap_id: The trap ID to check.
+    Returns:
+        bool: True if the trap is uninstalled, False otherwise.
+    """
+    return not is_trap_installed(duthost, trap_id)
+
+
+def get_trap_hw_status(duthost, namespace):
+    """
+    Retrieves the hw_status for traps from the STATE_DB.
+    Args:
+        dut (SonicHost): The target device
+    Returns:
+        dict: A dictionary mapping trap IDs to their hw_status.
+    """
+    if namespace is None:
+        state_db_cmd = "sonic-db-cli STATE_DB KEYS 'COPP_TRAP_TABLE|*'"
+        trap_data_cmd = "sonic-db-cli STATE_DB HGETALL "
+    else:
+        state_db_cmd = "sonic-db-cli -n {} STATE_DB KEYS 'COPP_TRAP_TABLE|*'".format(namespace.namespace)
+        trap_data_cmd = "sonic-db-cli -n {} STATE_DB HGETALL ".format(namespace.namespace)
+    state_db_data = duthost.shell(state_db_cmd)["stdout"]
+    state_db_data = state_db_data.splitlines()
+    hw_status = {}
+
+    for key in state_db_data:
+        trap_id = key.split("|")[-1]
+        trap_data = duthost.shell(trap_data_cmd + f"'{key}'")["stdout"]
+        trap_data_dict = ast.literal_eval(trap_data)
+        hw_status[trap_id] = trap_data_dict.get("hw_status", "not-installed")
+
+    return hw_status
+
+
+def get_random_copp_trap_config(duthost):
+    """
+    Retrieves a random CoPP trap config from /etc/sonic/copp_cfg.json on the DUT.
+    Returns the trap ID, its group, and related config details from COPP_TRAP and COPP_GROUP sections
+    Args:
+        duthost (SonicHost): The target device.
+    Returns:
+        tuple: A tuple containing the following elements:
+            - str: The first trap ID associated with the selected trap.
+            - str: The trap group associated with the selected trap.
+            - dict: The configuration details of the selected trap group from the `COPP_GROUP` section.
+    """
+
+    copp_cfg = json.loads(duthost.shell("cat /etc/sonic/copp_cfg.json")["stdout"])
+
+    # Get all traps from COPP_TRAP
+    copp_trap_cfg = copp_cfg.get("COPP_TRAP", {})
+    traps = list(copp_trap_cfg.keys())
+    assert traps, "No traps found in copp_cfg.json"
+
+    # Randomly select one trap
+    selected_trap = random.choice(traps)
+    trap_data = copp_cfg["COPP_TRAP"][selected_trap]
+    trap_ids = trap_data.get("trap_ids", "").split(",")
+    trap_group = trap_data.get("trap_group", "")
+    return trap_ids[0], trap_group, copp_cfg["COPP_GROUP"][trap_group]
+
+
+def get_feature_name_from_trap_id(duthost, trap_id):
+    """
+    Get the feature name corresponding to the given trap ID.
+    Args:
+        duthost (SonicHost): The target device.
+        trap_id (str): The trap ID to look up.
+    Returns:
+        bool: True if the trap ID is always enabled, False otherwise.
+        str: The feature name associated with the trap ID.
+    """
+
+    copp_cfg = json.loads(duthost.shell("cat /etc/sonic/copp_cfg.json")["stdout"])
+    copp_trap_cfg = copp_cfg.get("COPP_TRAP", {})
+
+    for feature_name, feature_data in copp_trap_cfg.items():
+        trap_ids = feature_data.get("trap_ids", "").split(",")
+        if trap_id in trap_ids:
+            always_enabled = feature_data.get("always_enabled", "false")
+            return always_enabled.lower() == "true", feature_name if always_enabled.lower() == "true" else feature_name
+
+    return False, None
